+++ tests/resources/views/components/attribute_passing_root.blade.php
@@ -1,0 +1,2 @@
+Root: <x-attribute_passing_child {{ $attributes }} />
+After Child: {{ $attributes }}